# packages/api-types/src/model/upscaleStartRequestDto.ts
/**
 * Generated by orval v8.15.0 🍺
 * Do not edit manually.
 * JellyfinSuite Plugin API
 * OpenAPI spec version: 1.0.0
 */

export interface UpscaleStartRequestDto {
  resultPath: string;
  scale: number;
  modelScale: number;
  modelStyle: string;
  faceRestore: boolean;
  /** @nullable */
  deviceId?: string | null;
}